Fixing use of paragraph and subparagraph
There should be no hard-coded numbering in documents like this, as
it will break as soon as sectioning elements are changed.  This
commit makes LaTeX enter the numbering itself.  While I'm at it, I'm also
changing the spacing of paragraph and subparagraph such that no
explicit markup is necessary any more.

Disclaimer: I've not checked how this comes out in HTML.  This would need to
be done at REC upload time.